+++ src/sql/mod.rs
@@ -1,3 +1,5 @@
+use std::fmt::Display;
+
 use crate::{
     error::{Error, Result},
     row::{self, Row},
@@ -11,7 +13,7 @@ mod ser;
 
 #[derive(Clone)]
 pub(crate) enum SqlBuilder {
-    InProgress { parts: Vec<Part>, size: usize },
+    InProgress(Vec<Part>),
     Failed(String),
 }
 
@@ -26,7 +28,6 @@ impl SqlBuilder {
     pub(crate) fn new(template: &str) -> Self {
         let mut iter = template.split('?');
         let prefix = String::from(iter.next().unwrap());
-        let mut size = prefix.len();
         let mut parts = vec![Part::Text(prefix)];
 
         for s in iter {
@@ -38,70 +39,83 @@ impl SqlBuilder {
                 s
             };
 
-            size += text.len();
             parts.push(Part::Text(text.into()));
         }
 
-        SqlBuilder::InProgress { parts, size }
+        SqlBuilder::InProgress(parts)
     }
 
     pub(crate) fn bind_arg(&mut self, value: impl Bind) {
-        if let Self::InProgress { parts, size } = self {
-            if let Some(part) = parts.iter_mut().find(|p| matches!(p, Part::Arg)) {
-                let mut s = String::new();
+        let Self::InProgress(parts) = self else {
+            return;
+        };
 
-                if let Err(err) = value.write(&mut s) {
-                    *self = SqlBuilder::Failed(err);
-                    return;
-                }
+        if let Some(part) = parts.iter_mut().find(|p| matches!(p, Part::Arg)) {
+            let mut s = String::new();
 
-                *size += s.len();
-                *part = Part::Text(s);
-            } else {
-                panic!("all query arguments are already bound");
+            if let Err(err) = value.write(&mut s) {
+                return self.error(format_args!("invalid argument: {err}"));
             }
+
+            *part = Part::Text(s);
+        } else {
+            self.error("unexpected bind(), all arguments are already bound");
         }
     }
 
     pub(crate) fn bind_fields<T: Row>(&mut self) {
-        if let Self::InProgress { parts, size } = self {
-            if let Some(fields) = row::join_column_names::<T>() {
-                for part in parts.iter_mut().filter(|p| matches!(p, Part::Fields)) {
-                    *size += fields.len();
-                    *part = Part::Text(fields.clone());
-                }
+        let Self::InProgress(parts) = self else {
+            return;
+        };
+
+        if let Some(fields) = row::join_column_names::<T>() {
+            for part in parts.iter_mut().filter(|p| matches!(p, Part::Fields)) {
+                *part = Part::Text(fields.clone());
             }
+        } else if parts.iter().any(|p| matches!(p, Part::Fields)) {
+            self.error("argument ?fields cannot be used with non-struct row types");
         }
     }
 
     pub(crate) fn append(&mut self, suffix: &str) {
-        if let Self::InProgress { parts, size } = self {
-            if let Some(Part::Text(text)) = parts.last_mut() {
-                *size += suffix.len();
-                text.push_str(suffix);
-            } else {
-                unreachable!();
-            }
+        let Self::InProgress(parts) = self else {
+            return;
+        };
+
+        if let Some(Part::Text(text)) = parts.last_mut() {
+            text.push_str(suffix);
+        } else {
+            // Do nothing, it will fail in `finish()`.
         }
     }
 
-    pub(crate) fn finish(self) -> Result<String> {
-        match self {
-            Self::InProgress { parts, size } => {
-                Ok(parts
-                    .into_iter()
-                    .fold(String::with_capacity(size), |mut res, part| {
-                        if let Part::Text(text) = part {
-                            debug_assert!(res.len() + text.len() <= res.capacity());
-                            res.push_str(&text);
-                            res
-                        } else {
-                            panic!("unbound query argument ? or ?fields");
-                        }
-                    }))
+    pub(crate) fn finish(mut self) -> Result<String> {
+        let mut sql = String::new();
+
+        if let Self::InProgress(parts) = &self {
+            for part in parts {
+                match part {
+                    Part::Text(text) => sql.push_str(text),
+                    Part::Arg => {
+                        self.error("unbound query argument");
+                        break;
+                    }
+                    Part::Fields => {
+                        self.error("unbound query argument ?fields");
+                        break;
+                    }
+                }
             }
+        }
+
+        match self {
+            Self::InProgress(_) => Ok(sql),
             Self::Failed(err) => Err(Error::InvalidParams(err.into())),
         }
+    }
+
+    fn error(&mut self, err: impl Display) {
+        *self = Self::Failed(format!("invalid SQL: {err}"));
     }
 }
 
@@ -120,8 +134,12 @@ mod tests {
         b: u32,
     }
 
+    #[allow(unused)]
+    #[derive(Row)]
+    struct Unnamed(u32, u32);
+
     #[test]
-    fn it_builds_sql_with_bound_args() {
+    fn bound_args() {
         let mut sql = SqlBuilder::new("SELECT ?fields FROM test WHERE a = ? AND b < ?");
         sql.bind_arg("foo");
         sql.bind_arg(42);
@@ -133,7 +151,7 @@ mod tests {
     }
 
     #[test]
-    fn it_builds_sql_with_in_clause() {
+    fn in_clause() {
         fn t(arg: &[&str], expected: &str) {
             let mut sql = SqlBuilder::new("SELECT ?fields FROM test WHERE a IN ?");
             sql.bind_arg(arg);
@@ -156,7 +174,7 @@ mod tests {
 
     // See #18.
     #[test]
-    fn it_builds_sql_with_question_marks_inside() {
+    fn question_marks_inside() {
         let mut sql = SqlBuilder::new("SELECT 1 FROM test WHERE a IN ? AND b = ?");
         sql.bind_arg(&["a?b", "c?"][..]);
         sql.bind_arg("a?");
@@ -164,5 +182,37 @@ mod tests {
             sql.finish().unwrap(),
             r"SELECT 1 FROM test WHERE a IN ['a?b','c?'] AND b = 'a?'"
         );
+    }
+
+    #[test]
+    fn failures() {
+        let mut sql = SqlBuilder::new("SELECT 1");
+        sql.bind_arg(42);
+        let err = sql.finish().unwrap_err();
+        assert!(err.to_string().contains("all arguments are already bound"));
+
+        let mut sql = SqlBuilder::new("SELECT ?");
+        sql.bind_arg(None::<u32>);
+        let err = sql.finish().unwrap_err();
+        assert!(err
+            .to_string()
+            .contains("invalid argument: serialize_none is unsupported"));
+
+        let mut sql = SqlBuilder::new("SELECT ?fields");
+        sql.bind_fields::<Unnamed>();
+        let err = sql.finish().unwrap_err();
+        assert!(err
+            .to_string()
+            .contains("argument ?fields cannot be used with non-struct row types"));
+
+        let mut sql = SqlBuilder::new("SELECT a FROM test WHERE b = ? AND c = ?");
+        sql.bind_arg(42);
+        let err = sql.finish().unwrap_err();
+        assert!(err.to_string().contains("unbound query argument"));
+
+        let mut sql = SqlBuilder::new("SELECT ?fields FROM test WHERE b = ?");
+        sql.bind_arg(42);
+        let err = sql.finish().unwrap_err();
+        assert!(err.to_string().contains("unbound query argument ?fields"));
     }
 }